Move DateDurationParseError to error module
