+++ components/calendar/src/error.rs
@@ -361,12 +361,113 @@ mod inner {
         Unimplemented,
         MonthCodeNotInCalendar,
     }
+
+    /// Errors that can occur when parsing an ISO 8601 date-only duration string.
+    ///
+    /// <div class="stab unstable">
+    /// 🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways,
+    /// including in SemVer minor releases. Do not use this type unless you are prepared for things to occasionally break.
+    ///
+    /// Graduation tracking issue: [issue #3964](https://github.com/unicode-org/icu4x/issues/3964).
+    /// </div>
+    #[derive(Debug, Copy, Clone, Eq, PartialEq)]
+    #[non_exhaustive]
+    pub enum DateDurationParseError {
+        /// The input does not follow the expected ISO 8601 date only duration structure.
+        ///
+        /// This error occurs when the duration string is incomplete,
+        /// or contains unexpected characters.
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// use icu::calendar::error::DateDurationParseError;
+        /// use icu::calendar::types::DateDuration;
+        ///
+        /// assert_eq!(DateDuration::try_from_str("P"),  Err(DateDurationParseError::InvalidStructure));
+        /// assert_eq!(DateDuration::try_from_str("P1"), Err(DateDurationParseError::InvalidStructure));
+        /// ```
+        InvalidStructure,
+
+        /// The duration contains a time component, which is not supported.
+        ///
+        /// Only date based units (`Y`, `M`, `W`, `D`) are supported. Any duration
+        /// containing a `T` time separator is rejected.
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// use icu::calendar::error::DateDurationParseError;
+        /// use icu::calendar::types::DateDuration;
+        ///
+        /// assert_eq!(DateDuration::try_from_str("PT5M"), Err(DateDurationParseError::TimeNotSupported));
+        /// assert_eq!(DateDuration::try_from_str("P1DT"), Err(DateDurationParseError::TimeNotSupported));
+        /// ```
+        TimeNotSupported,
+
+        /// A duration unit appeared without a number before it.
+        ///
+        /// For example, the string contains `Y`, `M`, `W`, or `D` without a
+        /// numeric value directly in front of it.
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// use icu::calendar::error::DateDurationParseError;
+        /// use icu::calendar::types::DateDuration;
+        ///
+        /// assert_eq!(DateDuration::try_from_str("PY"), Err(DateDurationParseError::MissingValue));
+        /// assert_eq!(DateDuration::try_from_str("PX1D"), Err(DateDurationParseError::MissingValue));
+        /// ```
+        MissingValue,
+
+        /// A duration unit was specified more than once.
+        ///
+        /// Each unit (`Y`, `M`, `W`, `D`) may appear at most once only.
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// use icu::calendar::error::DateDurationParseError;
+        /// use icu::calendar::types::DateDuration;
+        ///
+        /// assert_eq!(DateDuration::try_from_str("P1Y2Y"), Err(DateDurationParseError::DuplicateUnit));
+        /// assert_eq!(DateDuration::try_from_str("P1D1D"), Err(DateDurationParseError::DuplicateUnit));
+        /// ```
+        DuplicateUnit,
+
+        /// A numeric value exceeded or was more than the supported range.
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// use icu::calendar::error::DateDurationParseError;
+        /// use icu::calendar::types::DateDuration;
+        ///
+        /// assert_eq!(DateDuration::try_from_str("P4294967296Y"), Err(DateDurationParseError::NumberOverflow));
+        /// ```
+        NumberOverflow,
+
+        /// A duration starts with a `+` sign, which is not allowed.
+        ///
+        /// Only negative durations using a leading `-` are supported.
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// use icu::calendar::error::DateDurationParseError;
+        /// use icu::calendar::types::DateDuration;
+        ///
+        /// assert_eq!(DateDuration::try_from_str("+P1D"), Err(DateDurationParseError::PlusNotAllowed));
+        /// ```
+        PlusNotAllowed,
+    }
 }
 
 #[cfg(feature = "unstable")]
-pub use inner::EcmaReferenceYearError;
+pub use inner::{DateDurationParseError, EcmaReferenceYearError};
 #[cfg(not(feature = "unstable"))]
-pub(crate) use inner::EcmaReferenceYearError;
+pub(crate) use inner::{DateDurationParseError, EcmaReferenceYearError};
 
 impl From<EcmaReferenceYearError> for DateFromFieldsError {
     #[inline]
