+++ ffi/capi/tests/missing_apis.txt
@@ -17,6 +17,7 @@
 icu::calendar::Date::try_add_with_options#FnInStruct
 icu::calendar::Date::try_added_with_options#FnInStruct
 icu::calendar::Date::try_until_with_options#FnInStruct
+icu::calendar::error::DateDurationParseError#Enum
 icu::calendar::error::EcmaReferenceYearError#Enum
 icu::calendar::options::DateAddOptions#Struct
 icu::calendar::options::DateDifferenceOptions#Struct
@@ -27,7 +28,6 @@ icu::calendar::types::DateDuration::for_weeks#FnInStruct
 icu::calendar::types::DateDuration::for_years#FnInStruct
 icu::calendar::types::DateDuration::try_from_str#FnInStruct
 icu::calendar::types::DateDuration::try_from_utf8#FnInStruct
-icu::calendar::types::DateDurationParseError#Enum
 icu::calendar::types::DateDurationUnit#Enum
 icu::calendar::types::Month::leap#FnInStruct
 icu::calendar::types::Month::new#FnInStruct
